Moved listeners to redux kernel middleware, removed on disconnect
