+++ src/api/kernel.ts
@@ -62,18 +62,6 @@ export const connectToKernel = async (
       name: 'python3',
     });
 
-    // Attempt to shutdown the kernel on page exit
-    window.addEventListener(
-      'unload',
-      () => {
-        try {
-          kernel.shutdown();
-          syncSleep(1000);
-        } catch (error) {}
-      },
-      false
-    );
-
     console.log('Kernel started');
     return {
       success: true,
@@ -109,11 +97,3 @@ export const restart = (gatewayUri: string, kernel_id: string) =>
  */
 export const kill = (gatewayUri: string, kernel_id: string) =>
   fetch(`${gatewayUri}/api/kernels/${kernel_id}`, { method: 'DELETE' });
-
-/**
- * Sleep in-sync to delay close
- */
-export const syncSleep = (delay: number) => {
-  const start = new Date().getTime();
-  while (new Date().getTime() < start + delay);
-};
